refactor: split CLI command runners
